refactor: register page
